MR object

An MR object is a container for vectors and matrices of 2 or more GWAS data.

Usage

MR(
  exposure,
  outcome,
  harmonise_strictness = 2,
  correlation = NULL,
  verbose = TRUE
)

Arguments

exposure

a GWAS object or list of GWAS objects

outcome

a GWAS object

harmonise_strictness

an integer (1,2,3) corresponding to the TwoSampleMR harmonisation options of the same name.

correlation

a matrix, correlation matrix of signed R values between variants

verbose

a logical, print more information

Value

an S7 class genepi.utils::MR object

Slots

snps

character, variant ID

chr

character, chromosome identifier

bp

integer, base position

ea

character, effect allele

oa

character, other allele

eafx

numeric, exposure effect allele frequency

nx

integer, exposure total number of samples

ncasex

integer, exposure number of cases

bx

numeric, exposure effect size

bxse

numeric, exposure effect size standard error

px

numeric, exposure p-value

eafy

numeric, exposure effect allele frequency

ny

integer, exposure total number of samples

ncasey

integer, exposure number of cases

by

numeric, exposure effect size

byse

numeric, exposure effect size standard error

py

numeric, exposure p-value

exposure_id

character, the GWAS identifier

exposure

character, the GWAS exposure

outcome_id

character, the GWAS identifier

outcome

character, the GWAS outcome

group

integer, grouping variable used for plotting

index_snp

logical, whether the variant is an index variant (via clumping)

proxy_snp

character, the id of the proxy snp

ld_info

logical, whether there is LD information

info

data.frame, information about the loaded GWAS objects

correlation

matrix, a correlation matrix of signed R values between variants