Run collider bias analysis

Run all the analyses defined in the ColliderBias object x with all combinations of the listed parameters. Be careful with large numbers of parameters as this will quickly lead to very large numbers of combinations. e.g.

methods = c("slopehunter")

ip = c(0.05,0.001,0.00001)

pi0 = c(0.6, 0.65, 0.7)

sxy1 = c(1e-5, 1e-4, 1e-3)

...will lead to 27 separate analyses.

analyse(x, methods, ip, pi0, sxy1, bootstraps)

# S3 method for ColliderBias
analyse(
  x,
  methods = NULL,
  ip = NULL,
  pi0 = NULL,
  sxy1 = NULL,
  bootstraps = NULL
)

Arguments

x

a ColliderBias object

methods

a character vector, one or more valid collider bias analysis methods slopehunter, dudbridge, ivw_mr

ip

a numeric, range 0-1, the initial p-value by which to filter the incidence GWAS

pi0

a numeric, range 0-1, the initial weight / percentage of SNPs that affect incidence only

sxy1

a numeric, the initial (guess) of the covariance between incidence and progression BETAs

bootstraps

a whole-number numeric, the number of bootstrap samples to estimate the SE of the adjustment factor (can be zero)

Value

a ColliderBias object with updated parameter and results fields