Run collider bias assessment

Run collider bias assessment

collider_bias(
  x,
  bias_method = "dudbridge",
  r2 = 0.001,
  p1 = 5e-08,
  kb = 250,
  plink2 = genepi.utils::which_plink2(),
  plink_ref = genepi.utils::which_1000G_reference(build = "GRCh37"),
  ip = 0.001,
  pi0 = 0.6,
  sxy1 = 1e-05,
  bootstraps = 100,
  weighted = TRUE,
  method = "Simex",
  B = 1000,
  seed = 2023
)

Arguments

x

an object of class MR

bias_method

a character or character vector, one or more of c("dudbridge", "slopehunter", "mr_ivw", "mr_egger", "mr_weighted_median", "mr_weighted_mode")

r2

a numeric 0-1, r2 used for clumping - set all clumping params to NA to turn off

p1

a numeric 0-1, p1 used for clumping - set all clumping params to NA to turn off

kb

an integer, kb used for clumping - set all clumping params to NA to turn off

plink2

a path, the plink2 binary

plink_ref

a path, the reference genome pfile

ip

a numeric 0-1, threshold for removing incidence variants; see xp_thresh SlopeHunter::hunt()

pi0

a numeric 0-1, proportion of SNPs in the incidence only cluster; see init_pi SlopeHunter::hunt()

sxy1

a numeric, the covariance between incidence and progression Gip SNPs; see init_sigmaIP SlopeHunter::hunt()

bootstraps

an integer, number of bootstraps to estimate SE; see M SlopeHunter::hunt()

weighted

see weighted indexevent::indexevent()

method

see method indexevent::indexevent()

B

see B indexevent::indexevent()

seed

seed, for reproducibility