IVW MR collider bias method

The Inverse variance weighted Mendelian Randomisation method for assessing collider bias.

ivw_mr_collider_bias(
  gwas_i,
  gwas_p,
  ip = 0.001,
  merge = c(CHR = "CHR", BP = "BP"),
  ...
)

Arguments

gwas_i

a data.frame like object, clumped incidence GWAS summary statistics in standardise_gwas() format

gwas_p

a data.frame like object, progression GWAS summary statistics in standardise_gwas() format

ip

a numeric, range 0-1, the initial p-value by which to filter the incidence GWAS

merge

a named character vector c(gwas1_col=gwas2_col) of the columns to join on; e.g. c("CHR1"="CHR2","BP1"="BP2"); can be NULL, in which case the gwases must be data.tables and already have keys set

...

parameter sink, additional ignored parameters

Value

an S3 ColliderBiasResult object