QQ plot

QQ plot

Usage

qq_plot(
  gwas,
  pval_col = "p",
  colours = list(raw = "#2166AC"),
  title = NULL,
  subtitle = NULL,
  plot_corrected = FALSE,
  facet_grid_row_col = NULL,
  facet_grid_col_col = NULL,
  facet_nrow = NULL,
  facet_ncol = NULL
)

Arguments

gwas

a data.frame like object or valid file path

pval_col

the P value column

colours

a 2 element list of colour codes (1-the uncorrected points, 2-the GC corrected points)

title

a string, the title for the plot

subtitle

a string, the subtitle for the plot

plot_corrected

a logical, whether to apply and plot the lambda correction

facet_grid_row_col

a string, the column name in gwas by which to facet the plot (rows)

facet_grid_col_col

a string, the column name in gwas by which to facet the plot (cols)

facet_nrow

an integer, passed to facet_wrap, the number of rows to facet by (if only facet_grid_row_col is provided)

facet_ncol

an integer, passed to facet_wrap, the number of cols to facet by (if only facet_grid_col_col is provided)

Value

a ggplot