Liftover GWAS positions

Determine GWAS build and liftover to required build. This is the same function from the GwasDataImport package, the only difference being that you can specify the build rather than it trying to guess the build (which fails if you are trying to liftover small segments of the genome).

Usage

lift(
  gwas,
  from = "Hg19",
  to = "Hg38",
  snp_col = "snp",
  chr_col = "chr",
  pos_col = "pos",
  ea_col = "ea",
  oa_col = "oa",
  remove_duplicates = TRUE
)

Arguments

gwas

a data.table, or file path, chr, pos, snp name, effect allele, non-effect allele columns

from

which build to lift from, one of c("Hg18", "Hg19", "Hg38")

to

which build to lift over to, one of c("Hg18", "Hg19", "Hg38")

snp_col

Name of SNP column name. Optional. Uses less certain method of matching if not available

chr_col

Name of chromosome column name. Required

pos_col

Name of position column name. Required

ea_col

Name of effect allele column name. Optional. Might lead to duplicated rows if not presented

oa_col

Name of other allele column name. Optional. Might lead to duplicated rows if not presented

remove_duplicates

a logical, whether to remove duplicate IDs

Value

data.table with updated position columns

References

https://github.com/MRCIEU/GwasDataImport